+++ frontend/src/components/Tabs/index.tsx
@@ -1,0 +1,48 @@
+import * as stylex from '@stylexjs/stylex'
+import { s } from './style'
+import { useState } from 'react'
+
+export const Tabs = (props: {
+    tabs: string[]
+    children?: React.ReactElement[]
+    onChange?: (tab: typeof props['tabs'][number]) => unknown
+}) => {
+    type TTab = typeof props['tabs'][number]
+    const [currentTab, setCurrentTab] = useState<TTab>(props.tabs?.[0])
+
+    const onChangeTab = (tab: TTab) => {
+        setCurrentTab(tab)
+        props.onChange?.(tab)
+    }
+
+    return <div>
+        <div {...stylex.props(s.tabs)}>
+            {props.tabs.map(tab => (
+                <div 
+                    {...stylex.props(
+                        s.tab,
+                        currentTab === tab
+                            ? s.isActive
+                            : null
+                    )}
+                    onClick={() => onChangeTab(tab)}
+                >
+                    {tab}
+                </div>
+            ))}
+        </div>
+        <div>
+            {props.children?.map?.(element => ({
+                    ...element,
+                    props: {
+                        ...element.props,
+                        ...stylex.props(element.key === currentTab
+                            ? s.isActiveContent
+                            : s.tabContent
+                        ),
+                    }
+                }
+            ))}
+        </div>
+    </div>
+}
