+++ frontend/src/components/index.ts
@@ -23,3 +23,4 @@ export { Notification } from './Notification'
 export { PaginatedList } from './PaginatedList'
 export { FeaturedChannelList } from './FeaturedChannelList'
 export { AddFeaturedChannelForm } from './AddFeaturedChannelForm'
+export { Tabs } from './Tabs'
