+++ frontend/src/components/organisms/index.ts
@@ -1,0 +1,3 @@
+export { Channels } from './Channels'
+export { SuggestedChannels } from './SuggestedChannels'
+export { FeaturedChannels } from './FeaturedChannels'
