+++ frontend/src/pages/Admin/style.css
@@ -1,23 +1,9 @@
 .admin-page {
-  padding-bottom: 20px;
-}
-
-.channel-inputs {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 4px;
-}
-
-#form-container {
   display: flex;
   flex-direction: column;
   gap: 15px;
-}
-
-#languages {
-  display: grid;
-  grid-template-columns: auto auto;
+  padding-bottom: 20px;
+  width: 100%;
 }
 
 h2 {
