+++ frontend/src/utils/getFieldsWithUntrueValues.ts
@@ -1,0 +1,6 @@
+export const getFieldsWithUntrueValues = (object: Object) => {
+    return Object.entries(object).reduce((acc, [field, value]) => {
+        if (!value) acc.push(field)
+        return acc
+      }, [] as string[])
+}
