+++ frontend/src/utils/index.ts
@@ -2,3 +2,4 @@ export { isPortraitOrientation } from './isPortraitOrientation'
 export { delay } from './delay'
 export { getUrl } from './getUrl'
 export { getTgChannelName } from './getTgChannelName'
+export { getFieldsWithUntrueValues } from './getFieldsWithUntrueValues'
